+++ examples/anyscale-v2-existing-s3/outputs.tf
@@ -64,6 +64,7 @@ output "anyscale_register_command" {
     --s3-bucket-id ${split(":", var.existing_s3_bucket_arn)[5]} \
     --anyscale-iam-role-id ${module.aws_anyscale_v2_existing_s3.anyscale_iam_role_arn} \
     --instance-iam-role-id ${module.aws_anyscale_v2_existing_s3.anyscale_iam_role_cluster_node_arn} \
-    --efs-id ${module.aws_anyscale_v2_existing_s3.anyscale_efs_id}
+    --efs-id ${module.aws_anyscale_v2_existing_s3.anyscale_efs_id} \
+    --functional-verify workspace
   EOT
 }