+++ modules/aws-anyscale-iam/iam-policies-data.tf
@@ -51,7 +51,7 @@ data "aws_iam_policy_document" "iam_anyscale_crossacct_assumerole_policy" {
 }
 
 #Allow wildcard resources as these are locked down in other ways
-#tfsec:ignore:aws-iam-no-policy-wildcards
+#trivy:ignore:avd-aws-0342 trivy:ignore:avd-aws-0057
 data "aws_iam_policy_document" "iam_anyscale_steadystate_policy" {
   #checkov:skip=CKV_AWS_111:Write access required for these items
   #checkov:skip=CKV_AWS_356:Wildcards allowed for these items
@@ -239,6 +239,8 @@ data "aws_iam_policy_document" "iam_anyscale_steadystate_policy" {
     ]
     resources = ["*"]
   }
+
+  #trivy:ignore:avd-aws-0057
   statement {
     sid    = "ResourceManagementExtended"
     effect = "Allow"